alumnisocialnetworkapp/requirements.txt
asgiref==3.8.1
attrs==23.2.0
boto3==1.20.26
botocore==1.23.54
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.40.0
colorama==0.4.6
cryptography==42.0.7
decorator==4.4.2
dj-database-url==2.1.0
Django==5.0.6
django-anymail==10.2
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-cors-headers==3.14.0
django-crontab==0.7.1
django-debug-toolbar==4.4.2
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-oauth-toolkit==2.4.0
django-sslserver==0.22
django-storages==1.12.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
environs==10.0.0
exceptiongroup==1.2.0
gunicorn==21.2.0
h11==0.14.0
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
inflection==0.5.1
jmespath==0.10.0
jwcrypto==1.5.6
marshmallow==3.20.1
moviepy==1.0.3
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
proglog==0.1.10
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.6.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.0
s3transfer==0.5.2
selenium==4.19.0
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
stripe==7.9.0
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.7
wsproto==1.2.0